WELD-1269 Rename WeldListener to WeldInitialListener, but also keep WeldListener for backwards compatibility
