+++ docs/reference/src/main/docbook/pt-BR/ri-spi.po
@@ -436,7 +436,7 @@ msgstr "Servlet"
 
 #. Tag: para
 #, no-c-format
-msgid "If you are integrating Weld into a Servlet environment you must register <literal>org.jboss.weld.servlet.WeldListener</literal> as a Servlet listener, either automatically, or through user configuration, for each CDI application which uses Servlet."
+msgid "If you are integrating Weld into a Servlet environment you must register <literal>org.jboss.weld.servlet.WeldInitialListener</literal> as a Servlet listener, either automatically, or through user configuration, for each CDI application which uses Servlet."
 msgstr ""
 
 #. Tag: para
