+++ impl/src/main/java/org/jboss/weld/servlet/WeldListener.java
@@ -1,12 +1,6 @@
 /*
- * DO NOT ALTER OR REMOVE COPYRIGHT NOTICES OR THIS HEADER.
- *
- * Copyright 2009 Sun Microsystems, Inc. All rights reserved.
- *
- * Use is subject to license terms.
- *
  * JBoss, Home of Professional Open Source
- * Copyright 2008, Red Hat, Inc., and individual contributors
+ * Copyright 2013, Red Hat, Inc., and individual contributors
  * by the @authors tag. See the copyright.txt in the distribution for a
  * full listing of individual contributors.
  *
@@ -22,91 +16,12 @@
  */
 package org.jboss.weld.servlet;
 
-import static org.jboss.weld.logging.messages.ServletMessage.ONLY_HTTP_SERVLET_LIFECYCLE_DEFINED;
-import static org.jboss.weld.servlet.ConversationFilter.CONVERSATION_FILTER_REGISTERED;
-
-import javax.enterprise.inject.spi.CDI;
-import javax.inject.Inject;
-import javax.servlet.ServletContextEvent;
-import javax.servlet.ServletRequestEvent;
-import javax.servlet.http.HttpServletRequest;
-import javax.servlet.http.HttpSessionEvent;
-
-import org.jboss.weld.bean.builtin.BeanManagerProxy;
-import org.jboss.weld.exceptions.IllegalStateException;
-import org.jboss.weld.manager.BeanManagerImpl;
-import org.jboss.weld.servlet.api.helpers.AbstractServletListener;
-
 /**
- * The Weld listener
- * <p/>
- * Listens for context/session creation/destruction.
- * <p/>
- * Delegates work to the HttpContextLifecycle.
+ * @deprecated Renamed to WeldInitialListener
  *
- * @author Nicklas Karlsson
- * @author Dan Allen
- * @author Ales Justin
- * @author <a href="mailto:lincolnbaxter@gmail.com">Lincoln Baxter, III</a>
- * @author Jozef Hartinger
  * @author Marko Luksa
  */
-public class WeldListener extends AbstractServletListener {
+@Deprecated
+public class WeldListener extends WeldInitialListener {
 
-    @Inject
-    private BeanManagerImpl beanManager;
-    private HttpContextLifecycle lifecycle;
-
-    @Override
-    public void contextInitialized(ServletContextEvent sce) {
-        if (beanManager == null) {
-            // servlet containers may not be able to inject fields in a servlet listener
-            beanManager = BeanManagerProxy.unwrap(CDI.current().getBeanManager());
-        }
-        this.lifecycle = new HttpContextLifecycle(beanManager);
-        if (Boolean.valueOf(sce.getServletContext().getInitParameter(CONVERSATION_FILTER_REGISTERED))) {
-            this.lifecycle.setConversationActivationEnabled(false);
-        }
-        this.lifecycle.contextInitialized(sce.getServletContext());
-    }
-
-    @Override
-    public void contextDestroyed(ServletContextEvent sce) {
-        lifecycle.contextDestroyed(sce.getServletContext());
-    }
-
-    @Override
-    public void sessionCreated(HttpSessionEvent event) {
-        lifecycle.sessionCreated(event.getSession());
-    }
-
-    @Override
-    public void sessionDestroyed(HttpSessionEvent event) {
-        lifecycle.sessionDestroyed(event.getSession());
-    }
-
-    @Override
-    public void requestDestroyed(ServletRequestEvent event) {
-        if (event.getServletRequest() instanceof HttpServletRequest) {
-            lifecycle.requestDestroyed((HttpServletRequest) event.getServletRequest());
-        } else {
-            throw new IllegalStateException(ONLY_HTTP_SERVLET_LIFECYCLE_DEFINED);
-        }
-    }
-
-    @Override
-    public void requestInitialized(ServletRequestEvent event) {
-        if (lifecycle.isConversationActivationEnabled()) {
-            Object value = event.getServletContext().getAttribute(CONVERSATION_FILTER_REGISTERED);
-            if (Boolean.TRUE.equals(value)) {
-                this.lifecycle.setConversationActivationEnabled(false);
-            }
-        }
-
-        if (event.getServletRequest() instanceof HttpServletRequest) {
-            lifecycle.requestInitialized((HttpServletRequest) event.getServletRequest(), event.getServletContext());
-        } else {
-            throw new IllegalStateException(ONLY_HTTP_SERVLET_LIFECYCLE_DEFINED);
-        }
-    }
 }
